# docs/openapi/SCIMServer_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SCIMServer API",
    "version": "0.37.3",
    "description": "Complete SCIM 2.0 Server - 82 endpoints across 19 controllers.\n\nMulti-tenant endpoint architecture, OAuth 2.0, per-endpoint credentials, bulk operations, /Me endpoint, custom resource types, database browser, activity feed, per-endpoint logs, structured logging.\n\nAll SCIM paths are scoped under /endpoints/{endpointId}. Admin paths are under /admin.\nGlobal prefix: /scim (configurable via API_PREFIX). URL rewriting: /scim/v2/* → /scim/*.",
    "contact": { "name": "SCIMServer", "url": "https://github.com/pranems/SCIMServer" },
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
  },
  "servers": [
    { "url": "http://localhost:3000/scim", "description": "Local development (default)" },
    { "url": "http://localhost:6000/scim", "description": "Local (port 6000)" },
    { "url": "http://localhost:8080/scim", "description": "Docker" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "description": "Shared secret, per-endpoint credential, or OAuth JWT" },
      "oauthJwt": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "OAuth 2.0 JWT with sub claim - required for /Me" }
    },
    "schemas": {
...
      "PatchOp": {
        "type": "object",
        "required": ["schemas", "Operations"],
        "properties": {
          "schemas": { "type": "array", "items": { "type": "string" } },
          "Operations": { "type": "array", "maxItems": 1000, "items": { "type": "object", "required": ["op"], "properties": { "op": { "type": "string", "enum": ["add", "replace", "remove"] }, "path": { "type": "string" }, "value": {} } } }
        }
      },
      "SearchRequest": {
        "type": "object",
        "properties": {
          "schemas": { "type": "array", "items": { "type": "string" } },
          "filter": { "type": "string", "maxLength": 10000 },
          "startIndex": { "type": "integer", "minimum": 1 },
          "count": { "type": "integer", "minimum": 0, "maximum": 10000 },
          "sortBy": { "type": "string" },
          "sortOrder": { "type": "string", "enum": ["ascending", "descending"] },
          "attributes": { "type": "string", "description": "Comma-separated" },
          "excludedAttributes": { "type": "string", "description": "Comma-separated" }
        }
      },
...
      "Endpoint": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "displayName": { "type": "string" },
          "description": { "type": "string" },
          "active": { "type": "boolean" },
          "profile": { "type": "object", "description": "Full profile (view=full)" },
          "profileSummary": { "type": "object", "description": "Profile summary (view=summary)" },
          "scimBasePath": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "_links": { "type": "object", "properties": { "self": { "type": "string" }, "stats": { "type": "string" }, "credentials": { "type": "string" }, "scim": { "type": "string" } } }
        }
      },
...
      "UpdateEndpoint": { "type": "object", "properties": { "displayName": { "type": "string" }, "description": { "type": "string" }, "active": { "type": "boolean" }, "profile": { "type": "object" } } },
...
    },
    "parameters": {
      "endpointId": { "name": "endpointId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" }, "description": "Endpoint (tenant) UUID" },
      "resourceId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "SCIM resource ID" },
      "attributes": { "name": "attributes", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated attributes to include" },
      "excludedAttributes": { "name": "excludedAttributes", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated attributes to exclude" },
      "filter": { "name": "filter", "in": "query", "schema": { "type": "string" }, "description": "SCIM filter expression" },
      "startIndex": { "name": "startIndex", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "count": { "name": "count", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
      "sortBy": { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
      "sortOrder": { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["ascending", "descending"] } },
      "ifMatch": { "name": "If-Match", "in": "header", "schema": { "type": "string" }, "description": "ETag for conditional write" },
      "view": { "name": "view", "in": "query", "schema": { "type": "string", "enum": ["summary", "full"] } }
    }
  },
  "paths": {
...
    "/endpoints/{endpointId}/{resourceType}": {
      "post": { "tags": ["Custom Resources"], "summary": "Create custom resource", "operationId": "createCustomResource", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "name": "resourceType", "in": "path", "required": true, "schema": { "type": "string" } }, { "$ref": "#/components/parameters/attributes" }, { "$ref": "#/components/parameters/excludedAttributes" }], "requestBody": { "required": true, "content": { "application/scim+json": { "schema": { "type": "object" } } } }, "responses": { "201": { "description": "Created" }, "404": { "description": "Type not registered" } } },
      "get": { "tags": ["Custom Resources"], "summary": "List custom resources", "operationId": "listCustomResources", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "name": "resourceType", "in": "path", "required": true, "schema": { "type": "string" } }, { "$ref": "#/components/parameters/filter" }, { "$ref": "#/components/parameters/startIndex" }, { "$ref": "#/components/parameters/count" }, { "$ref": "#/components/parameters/sortBy" }, { "$ref": "#/components/parameters/sortOrder" }, { "$ref": "#/components/parameters/attributes" }, { "$ref": "#/components/parameters/excludedAttributes" }], "responses": { "200": { "description": "List" } } }
    },
    "/endpoints/{endpointId}/{resourceType}/.search": {
      "post": { "tags": ["Custom Resources"], "summary": "Search custom resources (POST)", "operationId": "searchCustomResources", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "name": "resourceType", "in": "path", "required": true, "schema": { "type": "string" } }], "requestBody": { "required": true, "content": { "application/scim+json": { "schema": { "$ref": "#/components/schemas/SearchRequest" } } } }, "responses": { "200": { "description": "Results" } } }
    },
    "/endpoints/{endpointId}/{resourceType}/{id}": {
      "get": { "tags": ["Custom Resources"], "summary": "Get custom resource", "operationId": "getCustomResource", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "name": "resourceType", "in": "path", "required": true, "schema": { "type": "string" } }, { "$ref": "#/components/parameters/resourceId" }, { "$ref": "#/components/parameters/attributes" }, { "$ref": "#/components/parameters/excludedAttributes" }], "responses": { "200": { "description": "Resource" }, "404": { "description": "Not found" } } },
      "put": { "tags": ["Custom Resources"], "summary": "Replace custom resource", "operationId": "replaceCustomResource", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "name": "resourceType", "in": "path", "required": true, "schema": { "type": "string" } }, { "$ref": "#/components/parameters/resourceId" }, { "$ref": "#/components/parameters/ifMatch" }], "requestBody": { "required": true, "content": { "application/scim+json": { "schema": { "type": "object" } } } }, "responses": { "200": { "description": "Replaced" }, "404": { "description": "Not found" } } },
      "patch": { "tags": ["Custom Resources"], "summary": "Patch custom resource", "operationId": "patchCustomResource", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "name": "resourceType", "in": "path", "required": true, "schema": { "type": "string" } }, { "$ref": "#/components/parameters/resourceId" }, { "$ref": "#/components/parameters/ifMatch" }], "requestBody": { "required": true, "content": { "application/scim+json": { "schema": { "$ref": "#/components/schemas/PatchOp" } } } }, "responses": { "200": { "description": "Patched" }, "404": { "description": "Not found" } } },
      "delete": { "tags": ["Custom Resources"], "summary": "Delete custom resource", "operationId": "deleteCustomResource", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "name": "resourceType", "in": "path", "required": true, "schema": { "type": "string" } }, { "$ref": "#/components/parameters/resourceId" }, { "$ref": "#/components/parameters/ifMatch" }], "responses": { "204": { "description": "Deleted" }, "404": { "description": "Not found" } } }
    },
...
    "/admin/endpoints/{endpointId}": {
      "get": { "tags": ["Admin - Endpoints"], "summary": "Get endpoint by ID", "operationId": "getEndpoint", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "$ref": "#/components/parameters/view" }], "responses": { "200": { "description": "Endpoint", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Endpoint" } } } }, "404": { "description": "Not found" } } },
      "patch": { "tags": ["Admin - Endpoints"], "summary": "Update endpoint (deep-merges settings)", "operationId": "updateEndpoint", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateEndpoint" } } } }, "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } } },
      "delete": { "tags": ["Admin - Endpoints"], "summary": "Delete endpoint (cascades all data)", "operationId": "deleteEndpoint", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }], "responses": { "204": { "description": "Deleted" }, "404": { "description": "Not found" } } }
    },
...
    "/admin/log-config/endpoint/{endpointId}/{level}": {
      "put": { "tags": ["Admin - Log Config"], "summary": "Set endpoint log level", "operationId": "setEndpointLogLevel", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }, { "name": "level", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Set" } } }
    },
    "/admin/log-config/endpoint/{endpointId}": {
      "delete": { "tags": ["Admin - Log Config"], "summary": "Clear endpoint log level override", "operationId": "clearEndpointLogLevel", "security": [{ "bearerAuth": [] }], "parameters": [{ "$ref": "#/components/parameters/endpointId" }], "responses": { "204": { "description": "Cleared" } } }
    },
...
  }
}